% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signer_operations.R
\name{signer_describe_signing_job}
\alias{signer_describe_signing_job}
\title{Returns information about a specific code signing job}
\usage{
signer_describe_signing_job(jobId)
}
\arguments{
\item{jobId}{[required] The ID of the signing job on input.}
}
\description{
Returns information about a specific code signing job. You specify the
job by using the \code{jobId} value that is returned by the StartSigningJob
operation.
}
\section{Request syntax}{
\preformatted{signer$describe_signing_job(
  jobId = "string"
)
}
}

\keyword{internal}
