% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDFSigner.R
\name{verify_pdf_signature}
\alias{verify_pdf_signature}
\title{Verifies the digital signature of a PDF file using pdfsig}
\usage{
verify_pdf_signature(pdf_file)
}
\arguments{
\item{pdf_file}{The path to the PDF file to be verified.}
}
\value{
A named list containing the parsed information about the digital signatures, or a message that no signatures were found.
}
\description{
This function uses the `pdfsig` command to check the digital signatures
of a PDF file.
}
\examples{
\dontrun{verify_pdf_signature("document.pdf")}
}
